Remove test question, fix typos, minor mistake
Using `super.method()` hasn't been explained yet on this page; it is explained in part 4. There is already a test question about it in part 4, so removed it here.
Also, the line `Engine engine = new Engine(...);` implied a fourth parameter was defined in the constructor, but this wasn't the case in the code above.